share the vision multi-head Attention used 6 times
The generic vision-tower attention (dims/num_heads multi-head attention) was
duplicated across gemma3, aya_vision, idefics3, lfm2_vl, paligemma and phi4_siglip
vision towers. Add mlx_vlm/models/attention.py with VisionAttention and repoint
those six vision modules to import it. Behavior-preserving pure extraction.
